changed data type to uint16_t

+++ linked_list.c
@@ -10,7 +10,7 @@ void list_init(Node** head, size_t size) {
 /// @brief inserts last in linked list
 /// @param head list head
 /// @param data data for the new node
-void list_insert(Node** head, int data) {
+void list_insert(Node** head, uint16_t data) {
     Node* new_node = mem_alloc(sizeof(Node));
     if(!new_node) return;
     new_node->data = data;
@@ -29,7 +29,7 @@ void list_insert(Node** head, int data) {
 /// @brief Inserts a node after prev_node
 /// @param prev_nodenode that will be before new node
 /// @param data data for the new node
-void list_insert_after(Node* prev_node, int data) {
+void list_insert_after(Node* prev_node, uint16_t data) {
     if (prev_node == NULL) return;
     Node* new_node = mem_alloc(sizeof(Node));
     if(!new_node) return;
@@ -42,7 +42,7 @@ void list_insert_after(Node* prev_node, int data) {
 /// @param head list head
 /// @param next_node node that will be after new node
 /// @param data data for the new node
-void list_insert_before(Node** head, Node* next_node, int data) {
+void list_insert_before(Node** head, Node* next_node, uint16_t data) {
     if (*head == NULL) return;  // ERROR
     Node* walker = *head;
     if (next_node == *head){
@@ -69,7 +69,7 @@ void list_insert_before(Node** head, Node* next_node, int data) {
 /// @brief deletes the Node with data
 /// @param head list head
 /// @param data
-void list_delete(Node** head, int data) {
+void list_delete(Node** head, uint16_t data) {
     if (*head == NULL) return;
     if ((*head)->data == data) {
         Node* temp = *head;
@@ -91,7 +91,7 @@ void list_delete(Node** head, int data) {
 /// @param head list head
 /// @param data value to search for
 /// @return Node* or NULL if node not found
-Node* list_search(Node** head, int data) {
+Node* list_search(Node** head, uint16_t data) {
     Node* walker = *head;
     while (walker != NULL) {
         if(walker->data == data) return walker;
